Add analyze command
