+++ cal_ratio_trainer/utils.py
@@ -1,6 +1,8 @@
 import argparse
+from collections import defaultdict
+import json
 from pathlib import Path
-from typing import Generator, Tuple, Union
+from typing import Generator, Optional, Tuple, Union
 
 
 def as_bare_type(t: type) -> Union[type, None]:
@@ -104,3 +106,95 @@ def apply_config_args(config_class: type, config, args):
             setattr(r, prop, getattr(args, prop))
 
     return r
+
+
+def find_training_result(
+    model_to_do: str, continue_from: Optional[int] = None, base_dir: Path = Path(".")
+) -> Path:
+    # If it doesn't exist, do our best.
+    model_dir = base_dir / "training_results" / model_to_do
+    if not model_dir.exists():
+        if continue_from is not None:
+            raise ValueError(
+                f"Directory {model_dir} does not exist. Cannot continue from the "
+                f"{continue_from} run it."
+            )
+        else:
+            return model_dir / "00000"
+
+    # The model dir exists. Lets see hwo well we can do here.
+    # If we have been given a continue_from directory, make sure it exists
+    # and then use that.
+    if continue_from is not None:
+        if continue_from < 0:
+            # get a sorted list of all sub-directories, and then take the one
+            # one back from the end.
+            try:
+                run_number = sorted(
+                    int(md.name) for md in model_dir.iterdir() if md.name.isdigit()
+                )[continue_from]
+                return model_dir / f"{run_number:05d}"
+            except IndexError:
+                raise ValueError(f"No runs in {model_dir} to continue from.")
+        else:
+            run_dir = model_dir / f"{continue_from:05d}"
+            if not run_dir.exists():
+                raise ValueError(
+                    f"Directory {run_dir} does not exist. Cannot continue from it."
+                )
+            return run_dir
+    # Ok - no continue_from listed - so we need to make up a new directory.
+    try:
+        biggest_run_number = max(
+            int(item.name)
+            for item in model_dir.iterdir()
+            if (item.is_dir() and item.name.isdigit())
+        )
+        biggest_run_number += 1
+    except ValueError:
+        biggest_run_number = 0
+
+    return model_dir / f"{biggest_run_number:05d}"
+
+
+class HistoryTracker:
+    "Keep together all the arrays we want to track per-epoch"
+
+    def __init__(self, file: Optional[Path] = None):
+        self._cache = defaultdict(list)
+
+        if file is not None:
+            self.load(file)
+
+    def __str__(self):
+        return f"HistoryTracker(epochs={len(self)}, items={list(self._cache.keys())})"
+
+    def __getattr__(self, name):
+        "Return the list for the requested tracking name"
+        return self._cache[name]
+
+    def __len__(self):
+        if len(self._cache) == 0:
+            return 0
+
+        else:
+            # Return the max length of all the lists
+            return max(len(item) for item in self._cache.values())
+
+    def save(self, filename: Path):
+        "Save the history to a file"
+
+        if filename.suffix != ".json":
+            filename = filename.with_suffix(".json")
+
+        with filename.open("w") as f:
+            json.dump(self._cache, f)
+
+    def load(self, filename: Path):
+        "Load history from a file"
+
+        if filename.suffix != ".json":
+            filename = filename.with_suffix(".json")
+
+        with filename.open("r") as f:
+            self._cache = json.load(f)
